Move remote config into wallet module
