--- app/src/main/java/com/algorand/android/koin/KoinHiltModule.kt
@@ -1,36 +1,0 @@
-/*
- * Copyright 2022 Pera Wallet, LDA
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at http://www.apache.org/licenses/LICENSE-2.0
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License
- */
-
-package com.algorand.android.koin
-
-import com.algorand.common.remoteconfig.domain.usecase.InitializeOperationalToggles
-import com.algorand.common.remoteconfig.domain.usecase.IsFeatureToggleEnabled
-import dagger.Module
-import dagger.Provides
-import dagger.hilt.InstallIn
-import dagger.hilt.components.SingletonComponent
-import org.koin.java.KoinJavaComponent.getKoin
-
-@Module
-@InstallIn(SingletonComponent::class)
-object KoinHiltModule {
-
-    @Provides
-    fun provideInitializeOperationalToggles(): InitializeOperationalToggles {
-        return getKoin().get()
-    }
-
-    @Provides
-    fun provideIsFeatureToggleEnabled(): IsFeatureToggleEnabled {
-        return getKoin().get()
-    }
-}
