+++ app/src/main/java/com/algorand/android/koin/KoinInitializer.kt
@@ -14,7 +14,6 @@ package com.algorand.android.koin
 
 import com.algorand.android.BuildConfig
 import com.algorand.android.core.PeraApp
-import com.algorand.common.remoteconfig.di.remoteConfigModule
 import org.koin.android.ext.koin.androidContext
 import org.koin.android.ext.koin.androidLogger
 import org.koin.core.context.startKoin
@@ -27,7 +26,6 @@ object KoinInitializer {
                 androidLogger()
             }
             androidContext(app)
-            modules(remoteConfigModule)
         }
     }
 }
