+++ common-sdk/build.gradle.kts
@@ -48,6 +48,9 @@ dependencies {
     implementation(libs.core.ktx)
     implementation(libs.appcompat)
     implementation(libs.material)
+    implementation(libs.firebase.analytics)
+    implementation(libs.firebase.config)
+    implementation(project.dependencies.platform(libs.firebase.bom))
 
     ksp(libs.dagger.hilt.compiler)
     ksp(libs.androidx.hilt.compiler)
